+++ db/schema.rb
@@ -22,4 +22,19 @@ ActiveRecord::Schema.define do
     t.datetime "created_at", :null => false
     t.datetime "updated_at", :null => false
   end
+
+  create_table "products", :force => true do |t|
+    t.string "name"
+    t.string "slug"
+    t.datetime "created_at", :null => false
+    t.datetime "updated_at", :null => false
+  end
+
+  create_table "tree_items", :force => true do |t|
+    t.string "name"
+    t.string "ancestry"
+    t.string "slug"
+    t.datetime "created_at", :null => false
+    t.datetime "updated_at", :null => false
+  end
 end