+++ lib/slugable/has_slug.rb
@@ -13,10 +13,12 @@ module Slugable
     # has_slug :from => :name, :to => :slug   # generate to_slug
     #
     def has_slug(options={})
-      defaults = {:from => :name, :to => :slug}
+      defaults = {:from => :name, :to => :slug, :formatter => :parameterize, :cache_tree => true}
       options.reverse_merge!(defaults)
       from = options.delete(:from)
       to = options.delete(:to)
+      formatter = options.delete(:formatter)
+      cache_tree = options.delete(:cache_tree)
       before_save :"fill_slug_from_#{from}_to_#{to}", :"format_slug_from_#{from}_to_#{to}"
       after_save :"update_my_#{to}_cache"
 
@@ -39,7 +41,7 @@ module Slugable
       # end
       code =<<-method
         def format_slug_from_#{from}_to_#{to}
-          self.#{to} = #{to}.parameterize
+          self.#{to} = #{to}.send(:#{formatter})
         end
       method
       class_eval(code)
@@ -98,12 +100,12 @@ module Slugable
       #
       # def to_slug
       #  if respond_to?(:path_ids)
-      #    slugs = path.map{|record| record.send(:"slug")}.select{|i| i.size > 0 }
-      #    if slugs.empty?
-      #      ""
+      #    slugs = if true
+      #      path_ids.map{|id| self.class.cached_slug(id)}.select{|i| i.size > 0 }
       #    else
-      #      slug
+      #      path.map{|record| record.send(:"slug")}.select{|i| i.size > 0 }
       #    end
+      #    slugs.empty? ? "" : slugs
       #  else
       #    send(:slug)
       #  end
@@ -111,19 +113,79 @@ module Slugable
       code =<<-method
         def to_#{to}
           if respond_to?(:path_ids)
-            slugs = path.map{|record| record.send(:"#{to}")}.select{|i| i.size > 0 }
-            if slugs.empty?
-              ""
+            slugs = if #{cache_tree}
+              path_ids.map{|id| self.class.cached_#{to}(id)}.select{|i| i.size > 0 }
             else
-              slugs
+              path.map{|record| record.send(:"#{to}")}.select{|i| i.size > 0 }
             end
+            slugs.empty? ? "" : slugs
           else
             send(:#{to})
           end
         end
       method
       class_eval(code)
-    end
 
+
+      # generate this
+      #
+      # def to_slug_was
+      #  if respond_to?(:ancestry_was)
+      #    old_slugs = if true
+      #      ancestry_was.to_s.split("/").map { |ancestor_id| self.class.cached_slug(ancestor_id.to_i) }
+      #    else
+      #      ancestry_was.to_s.split("/").map { |ancestor_id| self.class.find(ancestor_id).send(:slug) }
+      #    end
+      #    old_slugs << send(:slug_was)
+      #  else
+      #    send(:slug_was)
+      #  end
+      # end
+      code =<<-method
+        def to_#{to}_was
+          if respond_to?(:ancestry_was)
+            old_slugs = if #{cache_tree}
+              ancestry_was.to_s.split("/").map { |ancestor_id| self.class.cached_#{to}(ancestor_id.to_i) }
+            else
+              ancestry_was.to_s.split("/").map { |ancestor_id| self.class.find(ancestor_id).send(:#{to}) }
+            end
+            old_slugs << send(:#{to}_was)
+          else
+            send(:#{to}_was)
+          end
+        end
+      method
+      class_eval(code)
+
+      # generate this
+      #
+      # def to_slug_will
+      #  if respond_to?(:ancestry)
+      #    old_slugs = if true
+      #      ancestry.to_s.split("/").map { |ancestor_id| self.class.cached_slug(ancestor_id.to_i) }
+      #    else
+      #      ancestry.to_s.split("/").map { |ancestor_id| self.class.find(ancestor_id).send(:slug) }
+      #    end
+      #    old_slugs << send(:slug)
+      #  else
+      #    send(:slug_was)
+      #  end
+      # end
+      code =<<-method
+          def to_#{to}_will
+            if respond_to?(:ancestry)
+              new_slugs = if #{cache_tree}
+                ancestry.to_s.split("/").map { |ancestor_id| self.class.cached_#{to}(ancestor_id.to_i) }
+              else
+                ancestry.to_s.split("/").map { |ancestor_id| self.class.find(ancestor_id).send(:#{to}) }
+              end
+              new_slugs << send(:#{to}).send(:#{formatter})
+            else
+              send(:#{to}).send(:#{formatter})
+            end
+          end
+      method
+      class_eval(code)
+    end
   end
 end